Move GLSLValue to GL.Shader.Core.
